code overhaul - functionality changes
save_file() function has been add. If a file already exists in the destination folder, user is prompt to rename the file and save it to destination.

Decided to work with pack() instead of grid() which made things a lot easier. Replaced a few LabelFrames with Frames. Now has a listbox() to display what the program has done.